fit: the strength dial reaches the one window that ignored it
An Atmosphere zone's per-channel recolour was shrunk into [0.85, 1.18] by two
file-local constants. Every other limit in the solve interpolates with the
panel's Strength; this one did not, so a user who raised Strength to 1.0
because the sky was not blue enough got exactly the window a user at 0.0 got.
That zone is the commonest reason the dial gets raised at all.

It is now `FitBudget::zone_gain`, a point on the same piecewise ladder as
`wb_gain`, `slope` and `hsl_band`, and `shrink_atmosphere_gains_in` takes the
window as an argument. The DEFAULT point is the old pair of constants, which
stay named at their definition and are asserted equal to the budget's default
in the shrink test, so the shipped recipe is byte-identical and the existing
byte-identity pin covers the new axis too.

The same section asked whether `fit_cells` could also lift the zone's two
evidence withholdings, the way it lifts the white balance's and the mixer's.
It cannot, and the reason is worth recording where the code is: `fit_zone_dials`
solves `color_gains` from the zone's MASK-WEIGHTED MEAN moments, and a cell
voucher restricted to that same mask reads cell MEANS over the same pixels.
The estimator's objective and the voucher's measurement are one quantity, so
the verdict is not independent evidence - measured on the reference pair, it
read 1.000 converged / 0.000 diverged at every site whose probe was not
already null, and lifting it would have flipped twelve pinned refusals,
including the zero-evidence hue-band gate, for 0.0005 of frame residual on one
tile. The white balance and the mixer are vouched by cells because a weighted
median of per-pixel log ratios and a per-band population are genuinely
independent of the cell grid. The zone's two stay withheld, by name, as before.

+++ src/fit_cells.rs
@@ -115,7 +115,7 @@ impl PairedCells {
         let cells = CELLS_X * CELLS_Y;
         let of_pixel: Vec<usize> = (0..n)
             .map(|i| {
-                let (x, y) = (i as usize % width as usize, i as usize / width as usize);
+                let (x, y) = (i % width as usize, i / width as usize);
                 (y * CELLS_Y / height as usize) * CELLS_X + x * CELLS_X / width as usize
             })
             .collect();
